+++ src/fundamentalTypes/fileType.js
@@ -3,7 +3,9 @@ import {
 } from '../vendor-imports.js';
 import {$e} from '../utils/templateUtils.js';
 import {schemaLabel} from '../utils/schemaMeta.js';
-import {visualize, getUserMedia, startScreenCapture} from '../utils/media.js';
+import {
+  visualize, getUserMedia, startScreenCapture, isRealSafari
+} from '../utils/media.js';
 import dialogs from '../utils/dialogs.js';
 
 /**
@@ -368,7 +370,7 @@ const fileType = {
             class: 'video',
             $on: {
               loadeddata () {
-                if (!(/iPad|iPhone|iPod|Safari/u).test(navigator.userAgent)) {
+                if (!isRealSafari()) {
                   URL.revokeObjectURL(objURL);
                 }
               }
@@ -1065,7 +1067,7 @@ const fileType = {
                   recordedMedia.src = url;
 
                   recordedMedia.addEventListener('loadeddata', () => {
-                    if (!(/iPad|iPhone|iPod|Safari/u).test(navigator.userAgent)) {
+                    if (!isRealSafari()) {
                       URL.revokeObjectURL(url);
                     }
                   });